Add coreapi to dev requirements. Move all dev urls into separate config.
