+++ kolibri/core/templates/kolibri/base.html
@@ -20,7 +20,7 @@
   {% endif %}
   {% if debug %}
     <script src="{% static 'rest_framework/js/coreapi-0.1.1.js' %}"></script>
-    <script src="{% url 'kolibri:api-docs:schema-js' %}"></script>
+    <script src="{% url 'api-docs:schema-js' %}"></script>
   {% endif %}
 </head>
 <body>
