+++ kolibri/deployment/default/settings/dev.py
@@ -11,3 +11,5 @@ REST_SWAGGER = True
 REDIRECT_WEBPACK = True
 
 INTERNAL_IPS = ['127.0.0.1']
+
+ROOT_URLCONF = 'kolibri.deployment.default.dev_urls'
